Extract common logic to abstract class, replace deprecated client methods
